move event to it's own module
